backend_requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
anthropic>=0.40.0
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=3.0.0
pdfplumber>=0.10.0
pypdf>=4.0.0
